+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-# [4.6.1](23/01/2025)
+# [4.7.0] (23/01/2025)
+## Added
+Color structures such as `Color32` and `ColorF` have now been added. \
+`Color32` represents an `rgba` value between 0 and 255. \
+`ColorF` represents a normalized `rgba` value between 0 and 1f.
+
+# [4.6.1] (23/01/2025)
 ## Fixed
 The formatting of the `Vector4D` structure in the `ToString()` method where the `w` axis was not displayed, instead the `z` axis was displayed.